최종 등록 코드 1차 완성

+++ src/main/java/com/tikitta/backend/dto/ShowUpdateRequest.java
@@ -25,7 +25,7 @@ public class ShowUpdateRequest {
     private Long locationId;
     private String locationName;
     private String seatType;
-    private Long seatCount;
+    private Integer seatCount;
     private ShowMessageInfo showMessage;
     private String status;
     private String reviewUrl;
